Refactored run_catalogue
